reverted SkyBlockTime class extraction
